split up the application
