+++ galactory/constants.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+# (c) 2022 Brian Scholer (@briantist)
+
+CONTENT_TYPE = {'Content-Type': 'application/json'}
+HTTP_OK = 200
+HTTP_NOT_FOUND = 404
+HTTP_INTERNAL_SERVER_ERROR = 500
